% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robotdown.R
\docType{package}
\name{robotdown}
\alias{robotdown}
\alias{robotdown-package}
\title{robotdown: A package for creating undergraduate, Masters, and PhD theses
using R Markdown}
\description{
robotdown: A package for creating undergraduate, Masters, and PhD theses
using R Markdown
}
